app/requirements.txt
psycopg2-binary
pymongo
redis
fastapi
uvicorn
jinja2
python-multipart
asyncpg